Improve ConsoleCancellationManager with process termination timeout (#17251)

+++ src/Aspire.Cli/Commands/BaseCommand.cs
@@ -58,7 +58,7 @@ internal abstract class BaseCommand : Command
             catch (NonInteractiveException)
             {
                 // Error messages have already been displayed by the interaction service.
-                result = CommandResult.Failure((int)CliExitCodes.MissingRequiredArgument);
+                result = CommandResult.Failure(CliExitCodes.MissingRequiredArgument);
             }
 
             var isErrorExitCode = result.ExitCode != CliExitCodes.Success;
